% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaSPDE.R
\name{GammaSPDE}
\alias{GammaSPDE}
\title{Calculate Gamma}
\usage{
GammaSPDE(n = 10000)
}
\arguments{
\item{n}{natrual number for the approximation of the infinite sum. Default is 10000.}
}
\description{
Approximation of the constant Gamma which results by non-negligible covariances of squared increments in the SPDE model.
}
\examples{
Gamma <- GammaSPDE()
}
\references{
Bibinger, M. and Trabs, M. (2017) Volatility estimation for stochastic PDEs using high-frequency observations
}
\seealso{
\link{estimateParametersSPDE}, \link{kappa.test}.
}
\keyword{Approximation}
\keyword{Gamma.}
\keyword{of}
