% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappatest.R
\name{kappaTest}
\alias{kappaTest}
\title{Test for the curvature patameter in a parabolic SPDE model}
\usage{
kappaTest(
  data,
  curvature = 0,
  alternative = "two.sided",
  conf.level = 0.95,
  method = "non-oracle",
  spatialDelta = 0.05,
  sigma = NA,
  theta2 = NA,
  sigma0_squared = NA,
  theta1 = NA
)
}
\arguments{
\item{data}{a numerical \code{NxM} matrix. \code{N} denotes the temporal and \code{M} the spatial resolution of the grid.}

\item{curvature}{hypothesized curvature of the random field. Default is 0 (no curvature). Alternatively you can enter both hypothesized parameters \code{theta1} and \code{theta2}.}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter. Default is \code{two.sided}.}

\item{conf.level}{confidence level for the returned confidence interval.}

\item{method}{a string indicating if a non-orcale or an oracle test is performed. Must either be \code{non-oracle} or \code{oracle}. When \code{orcale} is used, you must specify the normalized voltility by either enter \code{sigma0_squared} or the parameters \code{sigma} and \code{theta2}. Default is \code{non-oracle}.}

\item{spatialDelta}{a real number greater than zero and less than 1/2 for selecting only the data points which are delta away from the Dirichlet boundary condition. The default is 0.05.}

\item{...}{further arguments depending on the chosen estimation method. See \code{estiamtionmethod} or the examples below.}
}
\description{
Aymptotic oracle and non-oracle tests for the curvature parameter in SPDE models. Non-oracle curvature tests are based on a log-linear model, see references Bibinger, M. and Bossert, P. (2022).
}
\examples{
# Data
theta0 = 0
theta1 = 0
theta2 = 1
sigma = 0.5
numberSpatialPoints = 100
numberTemporalPoints = 10000
spde <- simulateSPDEmodel(theta0,theta1,theta2,sigma,numberSpatialPoints,numberTemporalPoints)


kappa.test(spde,method = "non-oracle",curvature = -1,alternative = "g")
}
\references{
Bibinger, M. and Bossert, P. (2022) Efficient parameter estimation for parabolic SPDEs based on a log-linear model for realized volatilities.
}
\seealso{
\link{simulateSPDEmodel}, \link{estimateParametersSPDE}.
}
\keyword{Curvature}
\keyword{SPDEs.}
\keyword{Tests}
\keyword{for}
