% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSPDESamples.R
\name{MCSPDESamples}
\alias{MCSPDESamples}
\title{Simulates multiple parabolic stochastic partial differential equations in one space dimension using Monte-Carlo repetitions}
\usage{
MCSPDESamples(
  reputations,
  theta0,
  theta1,
  theta2,
  sigma,
  numberSpatialPoints,
  numberTemporalPoints,
  L = 10,
  xi = function(x) {
     0 * x
 },
  cutoff = 10000,
  method = "replacement"
)
}
\arguments{
\item{reputations}{A natural number giving the number of the Monte-Carlo repetitions.}

\item{...}{further arguments see \link{simulateSPDEmodel}.}
}
\value{
A list of numeric \code{NxM} matrices.
}
\description{
Simulate Monte-Carlo samples of a SPDE model in one space dimension on a discrete grid with \code{N} temporal and \code{M} spatial points, where the grid points are equidistant within the unit square. Each reputation uses the "replacement"-method by default, see \link{simulateSPDEmodel}.
}
\examples{
reputations = 10
theta0 = 0
theta1 = 1
theta2 = 1
sigma = 0.5
numberSpatialPoints = 10
numberTemporalPoints = 1000
l <- MCSPDESamples(reputations,theta0,theta1,theta2,sigma,numberSpatialPoints,numberTemporalPoints)
}
\references{
Bibinger, M. and Bossert, P. (2022) Efficient parameter estimation for parabolic SPDEs based on a log-linear model for realized volatilities,
Hildebrand, F. (2020) On generating fully discrete samples of the stochastic heat equation on an interval.
}
\seealso{
\link{simulateSPDEmodel}.
}
\keyword{Monte-Calro}
\keyword{SPDE.}
\keyword{samples}
