% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParabolicSPDEs.R
\name{ParabolicSPDEs}
\alias{ParabolicSPDEs}
\title{Parabolic stochastic partial differential equations in one space dimension}
\usage{
ParabolicSPDEs()
}
\description{
This package includes the simulation of SPDE models, \link{simulateSPDEmodel} and \link{MCSPDESamples}, as well as a quick-plot function \link{plotSPDE} and a function for estimating the natural parameters of the SPDE models \link{estimateParametersSPDE}.
}
\seealso{
\link{simulateSPDEmodel}, \link{MCSPDESamples}, \link{plotSPDE}, \link{estimateParametersSPDE}.
}
