% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPDE.R
\name{plotSPDE}
\alias{plotSPDE}
\title{Quick-plot SPDE sample data}
\usage{
plotSPDE(spdeData)
}
\arguments{
\item{spdeData}{a numeric \code{NxM} matrix, where \code{N} denotes the temporal and \code{M} the spatial resolution of the equidistant grid.}
}
\value{
a plotly graphic of the SPDE model.
}
\description{
Plotting a SPDE solution field on a discrete grid using \link[plotly:plot_ly]{plotly::plot_ly}.
}
\examples{
theta0 = 0
theta1 = 1
theta2 = 1
sigma = 0.5
numberSpatialPoints = 10
numberTemporalPoints = 1000
spde <- simulateSPDEmodel(theta0,theta1,theta2,sigma,numberSpatialPoints,numberTemporalPoints)

plotSPDE(spde)
}
\seealso{
\link{simulateSPDEmodel}.
}
\keyword{Quick-plot}
\keyword{SPDE}
\keyword{a}
\keyword{model.}
\keyword{of}
\keyword{samples}
