% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSPDEmodel.R
\name{simulateSPDEmodel}
\alias{simulateSPDEmodel}
\title{Simulate a parabolic stochastic partial differential equations in one space dimension}
\usage{
simulateSPDEmodel(
  theta0,
  theta1,
  theta2,
  sigma,
  numberSpatialPoints,
  numberTemporalPoints,
  L = 10,
  xi = function(x) {
     0 * x
 },
  cutoff = 10000,
  method = "replacement"
)
}
\arguments{
\item{theta0}{a real number which controls the drift of the solution field.}

\item{theta1}{a real number which controls the curvature of the solution field.}

\item{theta2}{a real number greater than zero which reduces the noise level of parameter sigma and the curvature effect of parameter \code{theta1}.}

\item{sigma}{a real number greater than zero which controls the overall noise level of the solution field.}

\item{L}{a natural number indicating the replacement bound LM dependent on multiples of \code{M}. The default is \code{L=10}.}

\item{xi}{initial condition. The default is \code{xi = 0}. For general initial condition choose \code{method = "cutoff"}.}

\item{cutoff}{a natural number for the cut-off frequency of the Fourrier series. Only used when \code{method = "cutoff"}. The default is \code{cutoff=10000}.}

\item{method}{either \code{"replacement"} (Default) or \code{"cutoff"}. Note, that the replacement method only allows the initial condition to be zero. For general initial conditions choose \code{method="cutoff"}.}

\item{numberOfSpatialPoints}{number of equidistant spatial points M in \code{[0,1]}.}

\item{numberOfTemporalPoints}{number of equidistant temporal points N in \code{[0,1]}.}
}
\value{
A numeric \code{NxM} matrix.
}
\description{
Simulate a sample of a SPDE model in one space dimension on a discrete grid with \code{N}-temporal and \code{M}-spatial points, where the grid points are equidistant within the unit square. The initial condition is set to be zero. For simulating SPDE samples with a general initial condition use the build-in cut-off method. Note that this method results in a systematic bias and dramatically increases computational costs. The SPDE model is using the Dirichlet boundary condition.
}
\examples{
theta0 = 0
theta1 = 1
theta2 = 1
sigma = 0.5
numberSpatialPoints = 10
numberTemporalPoints = 1000
spde <- simulateSPDEmodel(theta0,theta1,theta2,sigma,numberSpatialPoints,numberTemporalPoints)
}
\references{
Bibinger, M. and Bossert, P. (2022) Efficient parameter estimation for parabolic SPDEs based on a log-linear model for realized volatilities,
Hildebrand, F. (2020) On generating fully discrete samples of the stochastic heat equation on an interval.
}
\seealso{
\link{MCSPDESamples}, \link{plotSPDE},\link{estimateParametersSPDE}.
}
\keyword{SPDE}
\keyword{Sample}
\keyword{of}
\keyword{one}
